lib/context: make token validation errors more meaningful
This patch adds a message field for the Krb5Error::InvalidToken variant,
making it possible to provide more context about token validation errors.

Also the validation of the mic and wrap token headers is now done by
building the expected header and comparing it to the received token
header, so we can now include the expected header in error messages.